Specify the external library we link to on Windows
As of recently it is no longer sufficient to just build the `libX.a`
with the cc crate in build.rs. Due to
https://github.com/rust-lang/rust/issues/65610 it is also necessary to
specify the fact of linkage in the source for it to work in all
scenarios. The most frustrating part is that it only fails when shared
libraries are in the equation, which is a comparatively rare use-case in
Rust-land.