Add back Windows cross-compilation support
This adds back support for Windows cross-compilation by using the GAS
files if the Microsoft assembler isn't found rather than unconditionally
trying to use it targeting the `msvc` environment.

This also adds a test job to compile and test a cross compilation
scenario. If we take out testing and instead just test building then
that would be a much simpler step since we wouldn't need wine nor the
Windows SDK, which are only required when building/running the cargo
test executables.

Resolves: #58