Migrate archs that support inline assembly to Rust
This really has a couple benefits – first, is that we have an
opportunity to simplify the code in some instances. But most
importantly, we no longer need to deal with external assemblers and
different syntax for each.

Eventually we might even be able to utilize them macros to reduce
duplication too!